Change silent-renew path check
When Angular SPA is hosted in IIS, URL Rewriting must be employed to
send all requests to the SPA, otherwise IIS will return 404 on all
non-default requests, such as /silent-renew.

When the rewrite is used, the path changes a little in the browser and
so we need to check the absUrl instead of the path value to check for an
silent-renew request.